Calirni2

+++ backend/src/main/java/backend/backend/entity/UserRoleId.java
@@ -15,7 +15,7 @@ import java.io.Serializable;
 @Embeddable
 public class UserRoleId implements Serializable {
 
-	private Integer userId;
-    private Long roleId;
+    private Integer userId;
+    private Integer roleId;
 
 }
